pyproject.toml
[tool.poetry]
name = "gurman-backend"
version = "0.1.0"
description = ""
authors = ["Viktor Zagoruyko <v.a.zagoruyko.2@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
psycopg2-binary = "^2.9.10"
django-environ = "^0.12.0"
django-simple-history = "^3.10.1"
faker = "^37.8.0"
djangorestframework = "^3.16.1"
django-viewflow = "^2.2.12"
python-json-logger = "^3.3.0"
rollbar = "^1.3.0"
celery = "^5.5.3"
redis = "^6.4.0"
django-celery-beat = "^2.8.1"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
factory-boy = "^3.3.3"
pytest-cov = "^7.0.0"
django-cors-headers = "^4.9.0"
gunicorn = "^23.0.0"
pillow = "^11.3.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
isort = "^5.13"
flake8 = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
